pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "moescraper"
version = "0.1.0"
description = "Anime picture scraper toolkit."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "luminolous" }]
dependencies = [
  "httpx>=0.27",
  "pydantic>=2.0",
  "tqdm>=4.66",
  "Pillow>=10.0",
]

[project.urls]
Repository = "https://github.com/luminolous/moescraper"
Issues = "https://github.com/luminolous/moescraper/issues"

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "ruff>=0.6",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]  # pycodestyle, pyflakes, isort

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

